+++ gems/kettle-jem/spec/thin_slice_spec.rb
@@ -9107,6 +9107,8 @@ RSpec.describe Kettle::Jem do
         "We run rubocop on the latest version of Ruby",
       )
       expect(template_report.fetch(:final_content)).to include('gem "rubocop-lts", "~> 22.2.0"')
+      expect(template_report.fetch(:final_content)).to include('gem "rubocop-lts-rspec", "~> 1.0", ">= 1.0.1"')
+      expect(template_report.fetch(:final_content)).not_to include('gem "rubocop-rspec", "~> 3.6"')
       expect(template_report.fetch(:final_content)).to include('gem "rubocop-ruby3_1", "~> 3.0.0"')
       expect(template_report.dig(:metadata, :template_tokens)).to include(
         "KJ|RUBOCOP_LTS_CONSTRAINT" => "~> 22.2.0",